pyproject.toml
[project]
name = "finesightbench"
version = "0.1.0"
description = "Fine-grained visual perception & reasoning benchmark for VLMs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
]
